% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altitudes.R
\name{altitudes}
\alias{altitudes}
\title{Calculate altitude angles for a dem and range of azimuths}
\usage{
altitudes(
  dem = "raster1.tif",
  demDir = "~/projects/INRAE/data/",
  outDir = "~/projects/INRAE/data/altitudes/RCPP/",
  azimuthStep = 1,
  azimuthMin = NULL,
  azimuthMax = NULL,
  targetResolution = 10,
  gridConvergence = 1.48045031092839,
  correctCurvature = FALSE,
  useCPP = TRUE
)
}
\description{
Calculates altitude angles for each cell of a dem and a range of sun azimuths above which the cell potentially receives direct sunlight
}
