% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMinAltitudes.R
\name{calculateMinAltitudes}
\alias{calculateMinAltitudes}
\title{Calculate minimum altitudes}
\usage{
calculateMinAltitudes(dem, azimuth_min, azimuth_max, settings)
}
\arguments{
\item{dem}{dem raster}

\item{azimuth_min}{lower bounds of azimuths to calculate}

\item{azimuth_max}{upper bounds of azimuths}

\item{settings}{a settings object}
}
\value{
list of min altitudes for range of azimuths
}
\description{
Calculates minimum altitudes for a DEM raster and a range of azimuths
}
