% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutMinAltitudes.R
\name{cutMinAltitudes}
\alias{cutMinAltitudes}
\title{Calculate minimum altitudes}
\usage{
cutMinAltitudes(
  altitudeDir,
  altitudeFilePattern,
  altitudeFilePlaceholder = "azi",
  outDir,
  azimuthMin,
  azimuthMax,
  azimuthStep,
  stripeWidth = 10000
)
}
\arguments{
\item{dem}{dem raster}

\item{azimuth_min}{lower bounds of azimuths to calculate}

\item{azimuth_max}{upper bounds of azimuths}

\item{settings}{a settings object}
}
\value{
list of min altitudes for range of azimuths
}
\description{
Calculates minimum altitudes for a DEM raster and a range of azimuths
}
