% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doShades.R
\name{doShades}
\alias{doShades}
\title{Calculate a shade raster for a dem and a given time (UTC)}
\usage{
doShades(altitude, altitudes)
}
\description{
Calculate a shade raster for a dem and a given time (UTC)
}
