% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasShade.R
\name{hasShade}
\alias{hasShade}
\title{Calculate shades for a dem and a given time (UTC)}
\usage{
hasShade(
  dem = "raster1.tif",
  demDir = "~/projects/INRAE/data/",
  timeUTC = "2022-08-20 9:45:00",
  altitudesDir = "~/projects/INRAE/data/altitudes/RCPP/",
  outDir = "~/projects/INRAE/data/shademaps/RCPP/",
  azimuthStep = 1,
  targetResolution = 10
)
}
\description{
Determines shades for each cell based on pre-calculated altitude raster files for a given time
}
