% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precalcAltitudeDistances.R
\name{precalcAltitudeDistances}
\alias{precalcAltitudeDistances}
\title{Calculate altitude angles for a dem and range of azimuths}
\usage{
precalcAltitudeDistances(
  dem = "dem2.tif",
  demDir = "~/projects/INRAE/data/",
  outDir = "~/projects/INRAE/data/altitudes/RCPP/",
  azimuthStep = 1,
  azimuthMin = NULL,
  azimuthMax = NULL,
  targetResolution = 12.5,
  gridConvergence = 0,
  correctCurvature = FALSE,
  sampleIncFactor = 1
)
}
\description{
Calculates altitude angles for each cell of a dem and a range of sun azimuths above which the cell potentially receives direct sunlight
}
