% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precalcAltitudes.R
\name{precalcAltitudes}
\alias{precalcAltitudes}
\title{Calculate altitude angles for a dem and range of azimuths}
\usage{
precalcAltitudes(
  dem = "dem2.tif",
  demDir = "~/projects/INRAE/data/",
  outDir = "~/projects/INRAE/data/altitudes/RCPP/",
  azimuthStep = 1,
  azimuthMin = NULL,
  azimuthMax = NULL,
  gridConvergence = 0,
  correctCurvature = FALSE,
  sampleIncFactor = 1,
  cutVertically = FALSE,
  stripeWidth = 10000,
  originalResolution
)
}
\description{
Calculates altitude angles for each cell of a dem and a range of sun azimuths above which the cell potentially receives direct sunlight
}
