% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadeForTimeAndLocation.R
\name{shadeForTimeAndLocation}
\alias{shadeForTimeAndLocation}
\title{Calculate shades for a dem and a given time (UTC)}
\usage{
shadeForTimeAndLocation(
  timeUTC = "2022-08-20 7:45:00",
  lat = 42.75,
  lon = 0.85,
  dem = "zi_dtm_625_complete_cutlim.tif",
  rasterDEM = NULL,
  demDir = "~/projects/INRAE/data/",
  altitudesDir = "~/projects/INRAE/data/altitudes/625m/",
  azimuthStep = 5,
  targetResolution = 12.5,
  cutVertically = TRUE,
  stripeWidth = 10000
)
}
\description{
Determines shades for each cell based on pre-calculated altitude raster files for a given time
}
