% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunlightDurationForTimeAndArea.R
\name{sunlightDurationForTimeAndArea}
\alias{sunlightDurationForTimeAndArea}
\title{Calculate shades for a dem and a given time (UTC)}
\usage{
sunlightDurationForTimeAndArea(
  timeStartUTC = "2022-08-20 0:00:00",
  timeStopUTC = "2022-08-20 23:59:00",
  timestep = 15,
  altitudesDir = "~/projects/INRAE/data/altitudes/RCPP/stripes/",
  altitudeFilePattern = "altitudes_azimuth-{azi}_res-50_inc-1-03_stripe-1.tif",
  altitudeFilePlaceholder = "azi",
  outDir = "~/projects/INRAE/data/altitudes/RCPP/stripes/duration/",
  outFilename = "duration.tif",
  azimuthStep = 2,
  azimuthMin = 56
)
}
\description{
Determines shades for each cell based on pre-calculated altitude raster files for a given time
}
