% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunlightDurationForTimeAndLocation.R
\name{sunlightDurationForTimeAndLocation}
\alias{sunlightDurationForTimeAndLocation}
\title{Calculate shades for a dem and a given time (UTC)}
\usage{
sunlightDurationForTimeAndLocation(
  timeStartUTC = "2022-08-20 0:00:00",
  timeStopUTC = "2022-08-20 23:59:00",
  timestep = 15,
  lat = 42.75,
  lon = 0.85,
  dem = "zi_dtm_625_complete_cutlim.tif",
  demDir = "~/projects/INRAE/data/",
  altitudesDir = "~/projects/INRAE/data/altitudes/625m/",
  azimuthStep = 5,
  targetResolution = 12.5,
  cutVertically = TRUE,
  stripeWidth = 10000
)
}
\description{
Determines shades for each cell based on pre-calculated altitude raster files for a given time
}
