% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMinAltitudes.R
\name{calculateMinAltitudes}
\alias{calculateMinAltitudes}
\title{Calculate minimum altitudes}
\usage{
calculateMinAltitudes(dem, azimuth_min, azimuth_max, settings)
}
\arguments{
\item{settings}{a settings object}

\item{demFilename}{filename of the dem}
}
\value{
list of horizons for range of azimuths
}
\description{
Calculates minimum altitudes for a DEM raster and a range of azimuths
}
